feat(settings): expose remaining connection defaults
Expose the external server authorization and resource-check settings in the dialog, and stop seeding recent styles with built-in presets by default. Also bump the plugin version and clean up the control-layer lint warning.

+++ ai_diffusion/model/control.py
@@ -66,7 +66,7 @@ class ControlLayer(QObject, ObservableProperties):
         self._strength = max(
             0,
             min(
-                int(round(settings.control_layer_strength * self.strength_multiplier)),
+                round(settings.control_layer_strength * self.strength_multiplier),
                 self.max_strength,
             ),
         )
